add rendering tests
