+++ .github/workflows/benchmarks.yml
@@ -22,6 +22,9 @@ jobs:
         with:
           bundler-cache: true
 
+      - name: Run tests
+        run: rake test
+
       - name: Run benchmarks
         run: bundle exec rake benchmark
