+++ Rakefile
@@ -1,4 +1,9 @@
 # frozen_string_literal: true
+task :test do
+  ruby './test/render_test.rb'
+end
+
+task :default => :test
 
 task :benchmark do
   ruby './benchmark.rb'
