+++ phlex/phlex_name_component.rb
@@ -7,9 +7,12 @@ class PhlexNameComponent < Phlex::HTML
 
   def view_template
     h1 { "hello #{@name}" }
+    plain "\n"
 
     50.times do
       render PhlexNestedNameComponent.new(name: @name)
     end
+
+    plain "\n"
   end
 end
