+++ phlex/phlex_nested_name_component.rb
@@ -6,6 +6,8 @@ class PhlexNestedNameComponent < Phlex::HTML
   end
 
   def view_template
+    plain "\n  "
     p { "nested hello #{@name}" }
+    plain "\n"
   end
 end
